+++ guillotina/catalog/catalog.py
@@ -45,17 +45,17 @@ class DefaultSearchUtility:
 
     async def search(self, context: IBaseObject, query: typing.Any):
         """
-        Raw search query, uses parser to transform query
+        Search query, uses parser to transform query
         """
         parsed_query = parse_query(context, query, self)
         container = find_container(context)
         if container is not None:
-            return await self.query(container, parsed_query)
+            return await self.search_raw(container, parsed_query)
         raise ContainerNotFound()
 
-    async def query(self, container: IContainer, query: typing.Any):
+    async def search_raw(self, container: IContainer, query: typing.Any):
         """
-        Search parsed query
+        Search raw query
         """
         return {"items": [], "items_total": 0}